LPS-97345 Add required path/query parameters and extract an interface from the UserResource so that the implementation from the scim-impl and another from the charon integration module share the same contract. Also, rename class to SCIMUserResourceImpl
