+++ modules/dxp/apps/scim/scim-api/src/main/resources/com/liferay/scim/resource/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
