+++ modules/dxp/apps/scim/source-formatter-suppressions.xml
@@ -2,6 +2,6 @@
 
 <suppressions>
 	<source-check>
-		<suppress checks="JavaLongLinesCheck" files="scim-impl/src/main/java/com/liferay/scim/internal/jaxrs/application/UserResource\.java" />
+		<suppress checks="JavaLongLinesCheck" files="scim-impl/src/main/java/com/liferay/scim/internal/resource/SCIMUserResourceImpl\.java" />
 	</source-check>
 </suppressions>
